Rename constants.ts to version.ts
