+++ app/index.tsx
@@ -12,11 +12,11 @@ import "@foxglove-studio/app/styles/global.scss";
 
 import App from "@foxglove-studio/app/App";
 import OsContextSingleton from "@foxglove-studio/app/OsContextSingleton";
-import { APP_VERSION } from "@foxglove-studio/app/constants";
 import installDevtoolsFormatters from "@foxglove-studio/app/util/installDevtoolsFormatters";
 import { initializeLogEvent } from "@foxglove-studio/app/util/logEvent";
 import overwriteFetch from "@foxglove-studio/app/util/overwriteFetch";
 import waitForFonts from "@foxglove-studio/app/util/waitForFonts";
+import { APP_VERSION } from "@foxglove-studio/app/version";
 import { Sockets } from "@foxglove/electron-socket/renderer";
 import Logger from "@foxglove/log";
 
@@ -32,7 +32,7 @@ if (
   initSentry({
     dsn: process.env.SENTRY_DSN,
     autoSessionTracking: true,
-    release: "studio@" + APP_VERSION,
+    release: `${process.env.SENTRY_PROJECT}@${APP_VERSION}`,
   });
 }
