+++ webpack.main.config.ts
@@ -72,6 +72,7 @@ export default (_: unknown, argv: WebpackArgv): Configuration => {
       }),
       new EnvironmentPlugin({
         SENTRY_DSN: process.env.SENTRY_DSN ?? null, // eslint-disable-line no-restricted-syntax
+        SENTRY_PROJECT: process.env.SENTRY_PROJECT ?? null, // eslint-disable-line no-restricted-syntax
         AMPLITUDE_API_KEY: process.env.AMPLITUDE_API_KEY ?? null, // eslint-disable-line no-restricted-syntax
       }),
       new ForkTsCheckerWebpackPlugin(),
