+++ webpack.preload.config.ts
@@ -46,6 +46,7 @@ export default (_: unknown, argv: WebpackArgv): Configuration => {
     plugins: [
       new EnvironmentPlugin({
         SENTRY_DSN: process.env.SENTRY_DSN ?? null, // eslint-disable-line no-restricted-syntax
+        SENTRY_PROJECT: process.env.SENTRY_PROJECT ?? null, // eslint-disable-line no-restricted-syntax
         AMPLITUDE_API_KEY: process.env.AMPLITUDE_API_KEY ?? null, // eslint-disable-line no-restricted-syntax
       }),
       new ForkTsCheckerWebpackPlugin(),
